OJ-2421: AWS Test of Step Function publish integration
When Step Function publish data for audit event in we don't
know if they work, Eventbridge lacks a storage mechanism.
The idea behind this PR is to temporarily adds test queues
as a target to existing rules and eventbridge. The test can then assert
on messages that land in the test queue. The data payload can be
inspected in the test to ensure the right items are sent.
Any changes can be caught of things change in future.