OJ-2556: Refactor event publishing in AuditEvent StepFunction
In order, to make testing the payload sent to TxMA similar to
to how the putEvent integration of the CRI step functions. It makes
sense to decouple publishing to TxMA but substitution a publish event
integration with a putEvent integration. Doing this would enable
targeting a test queue and verifying the shape of the payload going
to the TxMA Queue.

An alarm is need to for the rule that triggers the TxMA Queue