OJ-1851: Update Step Function Test
Remove eventBridge publishing testing from main Step Function flow
test.

Since the AWS tests are running almost and the same time. Collision
can occur.

To prevent this, use different sessionId across the tests.
This is most evident in the nino_issue_credential tests which
in addition to this also different bearer token values this
was key to get that test working