OJ-2230: Removed checking ttlValue to be positive
It is possible a past time for is required for some logic i.e a time has to be x minutes in the past
If that situation arises in the step function then the handler would be ready to process the request
From a time handler perspective - it doesn't know what context it's going to be used in.
This would ultimately be decided by the step function that uses it.