BAU: Minor refactor to state machine alarms
The alarms are transition to OK every morning due to MissingData being treated as missing. These alarms all trigger slack messages causing the channel to get spammed.
Refactored to treatMissingData as notBreaching.

Also, previously the alarms did not match the description, rather than 4 failures within an hour, it was atleast 1 failure in 4 different 5 minute periods within an hour. This would mean if for example, if 100 executions failed but only within a 5 minute period it would not trigger the alarm.
Refactored to make the alarm match the description.