OJ-3227: Query access token index
When using the QueryCommand on the Session Table we need to query using the partion key (sessionId), not the accessToken.

Refactored to use the access-token-index to initially retrieve the sessionId, and added retries for this given recent issues with GSIs. Then query the Session Table with the sessionId. This is the same as the existing State machine.

Note: TTL does not need checking on the access-token-index. When we query the Session table with the sessionId it will check the TTL.