OJ-3228: updated DatabasePopulator
The issue was that the TTL column in DDB accepts either a string OR a number,
but if you query with a string as your filter value you won't see the numeric
values and vice versa. We use numeric values for this column everywhere,
except the contract test setup which happened to be pushing a string,
and this was what caused the queries to fail. The difference between
the step function and the Lambda is that the Lambdas always check
the TTL column as part of the database query