OJ-3434 - Remove client_id expectation from integration tests
The clientId in the session table is not the same thing as
the client_id in the audit event. The former is the caller
of our specific CRI, whereas the latter is the caller of
One Login (ie, the RP). Leaving client_id unset allows the
TxMA event processing logic to enrich the event with the
client id from other events that correspond to the same
user journey.